Rename norwegian locale
Closes #854
