+++ src/Faker/Provider/nb_NO/Company.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Faker\Provider\no_NO;
+namespace Faker\Provider\nb_NO;
 
 class Company extends \Faker\Provider\Company
 {
